requirements.txt
discord.py
psycopg2-binary
bs4
lxml
requests
